iot_client/requirements.txt
awscrt==0.6.0
awsiotsdk==1.3.0
boto3==1.14.47
botocore==1.17.47
docutils==0.15.2
jmespath==0.10.0
python-dateutil==2.8.1
s3transfer==0.3.3
six==1.15.0
urllib3==1.25.10
